Moves component test to component, moves server actions next to page
Realized the create action should just be passed into the component. doing this made testing much easier. the modal tests now properly assert that actions are only being triggered if validation passes.
